+++ docs/CHANGELOG.md
@@ -3,3 +3,6 @@
 ### Initialize repository 0604019 
     * repository created
 
+### Initialize repository 0105019 
+    * Project 2 added (LM35 to LCD/IC2)
+